requirements.txt
Django>=3.2.6,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9.1,<2.10.0
requests>=2.26.0,<2.27.0
flake8>=3.9.2,<3.10.0
